[WIP] Refactors the quests' grammars' terminals to scriptable objects and to substitute types to polymorphism (done). Adds the corresponding SO to each content the quest corresponds to (doing). Changes the quests' parameters, that are passed to the generators, to better represent the game's contents (doing). Fixes the creation of each generator's parameters based on the quest line (doing). Rearranges some files to better organize the scripts and namespaces.

+++ Assets/Scripts/Game/LevelManager/Map.cs
@@ -119,7 +119,7 @@ namespace Game.LevelManager
                         {
                             if (actualRoom.IsLeafNode())
                             {
-                                treasure = UnityEngine.Random.Range(0, GameManager.instance.treasureSet.Items.Count);
+                                treasure = UnityEngine.Random.Range(0, (int)GameManager.instance.treasureSet.Items.Count);
                                 dungeonPartCode = DungeonPart.PartType.TREASURE_ROOM;
                             }
                         }
